requirements.txt
eventlet==0.29.0
python-socketio==4.6.0
redis==3.5.3
